qcacld-3.0: Check for target ready before sending peer create
When modem shutdown is triggered and HOST driver is in the
middle of a connection attempt, it doesnt check if the fw target
is ready. Instead the driver tries to join multiple AP with same
ssid and for each connection it tries to send peer add as part
of wma_set_link_state.

Add check to see if FW is down in wma_create_peer. Return
failure to LIM instead of sending add peer to firmware.

Change-Id: I4642c844bcc1d7f32888b0f89db5c11ee84c41fe
CRs-Fixed: 2267967